--- a/.github/workflows/ci-fast.yml
+++ b/.github/workflows/ci-fast.yml
@@ -132,7 +132,7 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
       with:
-        file: ./coverage/coverage.xml
+        files: ./coverage/coverage.xml
         flags: unittests
         name: codecov-umbrella
         fail_ci_if_error: false