--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -141,7 +141,7 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
       with:
-        file: ./coverage/coverage.xml
+        files: ./coverage/coverage.xml
         flags: unittests
         name: codecov-umbrella
         fail_ci_if_error: false