--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -151,6 +151,16 @@ jobs:
         go-version: ${{ env.GO_VERSION }}
         cache: true
 
+    - name: Install Wire tool
+      run: |
+        go install github.com/google/wire/cmd/wire@v0.6.0
+        echo "$HOME/go/bin" >> $GITHUB_PATH
+
+    - name: Generate Wire code
+      run: |
+        echo "Generating Wire dependency injection code..."
+        wire ./cmd/usercenter
+
     - name: Build for multiple platforms
       run: make build-all